player boards rework
