+++ src/components/overview/OverviewSettings.ts
@@ -18,8 +18,8 @@ export const OverviewSettings = Vue.component('overview-settings', {
   },
   template: `
         <div class="players-overview-settings">
-            <div class="setting-label">toggle tags view:</div>
             <div class="setting-button" v-on:click.prevent="toggleTagsView()">{{ getTagToggleLabel() }}</div>
+            <div class="setting-label">[ toggle tags view ]</div>
         </div>
     `,
 });
