+++ src/components/overview/PlayerInfo.ts
@@ -4,7 +4,19 @@ import {PlayerResources} from './PlayerResources';
 import {PlayerTags} from './PlayerTags';
 import {PlayerStatus} from './PlayerStatus';
 import {playerColorClass} from '../../utils/utils';
+import {mainAppSettings} from '../App';
+import {range} from '../../utils/utils';
+import {PlayerMixin} from '../PlayerMixin';
 
+const isPinned = (root: any, playerIndex: number): boolean => {
+  return (root as any).getVisibilityState('pinned_player_' + playerIndex);
+};
+const showPlayerData = (root: any, playerIndex: number) => {
+  (root as any).setVisibilityState('pinned_player_' + playerIndex, true);
+};
+export const hidePlayerData = (root: typeof mainAppSettings.methods, playerIndex: number) => {
+  root.setVisibilityState('pinned_player_' + playerIndex, false);
+};
 export const PlayerInfo = Vue.component('player-info', {
   props: {
     player: {
@@ -31,18 +43,10 @@ export const PlayerInfo = Vue.component('player-info', {
     'player-tags': PlayerTags,
     'player-status': PlayerStatus,
   },
+  mixins: [PlayerMixin],
   methods: {
     getClasses: function(): string {
       const classes = ['player-info'];
-      return classes.join(' ');
-    },
-    getNameAndIconClasses: function(): string {
-      const classes = ['name-and-icon'];
-      classes.push(playerColorClass(this.player.color, 'bg_transparent'));
-      return classes.join(' ');
-    },
-    getPlayerCorpClasses: function(): string {
-      const classes = ['player-corp'];
       classes.push(playerColorClass(this.player.color, 'bg_transparent'));
       return classes.join(' ');
     },
@@ -50,32 +54,71 @@ export const PlayerInfo = Vue.component('player-info', {
       const classes = ['player-status-and-res'];
       return classes.join(' ');
     },
-    getInfoBottomClasses: function(): string {
-      const classes = ['player-info-bottom'];
-      classes.push(playerColorClass(this.player.color, 'bg_transparent'));
-      return classes.join(' ');
-    },
     getIsActivePlayer: function(): boolean {
       return this.player.color === this.activePlayer.color;
+    },
+    pinPlayer: function() {
+      let hiddenPlayersIndexes: Array<Number> = [];
+      const playerPinned = isPinned(this.$root, this.playerIndex);
+
+      // if player is already pinned, add to hidden players (toggle)
+      hiddenPlayersIndexes = range(this.activePlayer.players.length - 1);
+      if (!playerPinned) {
+        showPlayerData(this.$root, this.playerIndex);
+        hiddenPlayersIndexes = hiddenPlayersIndexes.filter(
+          (index) => index !== this.playerIndex,
+        );
+      }
+      for (let i = 0; i < hiddenPlayersIndexes.length; i++) {
+        if (hiddenPlayersIndexes.includes(i)) {
+          hidePlayerData(this.$root as unknown as typeof mainAppSettings.methods, i);
+        }
+      }
+    },
+    buttonLabel: function(): string {
+      return isPinned(this.$root, this.playerIndex) ? 'hide' : 'show';
+    },
+    getNrPlayedCards: function(): number {
+      return (this.player.corporationCard !== undefined ? 1 : 0) + this.player.playedCards.length;
+    },
+    togglePlayerDetails: function() {
+      // for active player => scroll to cards UI
+      if (this.player.color === this.activePlayer.color) {
+        const el: HTMLElement = document.getElementsByClassName(
+          'preferences_icon--cards',
+        )[0] as HTMLElement;
+        el.click();
+
+        return;
+      }
+      // any other player show cards container and hide all other
+      this.pinPlayer();
     },
   },
   template: ` 
       <div :class="getClasses()">
-        <div class="player-info-top">
-            <div :class="getNameAndIconClasses()">
-              <div class="player-info-name">{{ player.name }}</div>
-              <div class="icon-first-player" v-if="firstForGen && activePlayer.players.length > 1">1st</div>
-            </div>
-            <div v-if="player.corporationCard !== undefined" :title="player.corporationCard.name" :class="getPlayerCorpClasses()">{{ player.corporationCard.name }}</div>
-            <div class="player-discounts-background" />
-        </div>
-        <div :class="getInfoBottomClasses()">
-          <div :class="getPlayerStatusAndResClasses()">
-            <player-status :player="player" :activePlayer="activePlayer" :firstForGen="firstForGen" v-trim-whitespace :actionLabel="actionLabel" :playerIndex="playerIndex"/>
-            <player-resources :player="player" v-trim-whitespace />
+        <div :class="getPlayerStatusAndResClasses()">
+        <div class="player-status">
+          <div class="player-info-details">
+            <div class="player-info-name">{{ player.name }}</div>
+            <div class="icon-first-player" v-if="firstForGen && activePlayer.players.length > 1">1st</div>
+            <div class="player-info-corp" v-if="player.corporationCard !== undefined" :title="player.corporationCard.name">{{ player.corporationCard.name }}</div>
           </div>
-          <player-tags :player="player" v-trim-whitespace :isActivePlayer="getIsActivePlayer()" :hideZeroTags="hideZeroTags" />
+          <player-status :player="player" :activePlayer="activePlayer" :firstForGen="firstForGen" v-trim-whitespace :actionLabel="actionLabel" :playerIndex="playerIndex"/>
         </div>
+          <player-resources :player="player" v-trim-whitespace />
+          <div class="player-played-cards">
+            <div class="player-played-cards-top">
+              <div class="played-cards-icons">
+                <div class="played-cards-icon hiding-card-button active"> {{ getCardsByType(player.playedCards, [getActiveCardType()]).length.toString() }} </div>
+                <div class="played-cards-icon hiding-card-button automated"> {{ getCardsByType(player.playedCards, [getAutomatedCardType()]).length.toString() }} </div>
+                <div class="played-cards-icon hiding-card-button event"> {{ getCardsByType(player.playedCards, [getEventCardType()]).length.toString() }} </div>
+              </div>
+            </div>
+            <Button class="played-cards-button" size="tiny" :onClick="togglePlayerDetails" :title="buttonLabel()" />
+          </div>
+        </div>
+        <player-tags :player="player" v-trim-whitespace :isActivePlayer="getIsActivePlayer()" :hideZeroTags="hideZeroTags" />
       </div>
     `,
 });
