+++ src/styles/board.less
@@ -294,16 +294,19 @@
 
 	.hide-tile-button {
 		position: absolute;
-    width: fit-content;
-    white-space: nowrap;
-    bottom: -560px;
-    text-transform: uppercase;
-    border-radius: 4px;
-    font-size: @font_size_small;
-    padding: 1px 5px;
-    color: black;
-    background: #666;
-    cursor: pointer;
+		bottom: -500px;
+		text-transform: uppercase;
+		font-size: 12px;
+		color: #ffcc64;
+		background: #303030;
+		clip-path: polygon(50% 0, 100% 25%, 100% 75%, 50% 100%, 0 75%, 0 25%);
+		cursor: pointer;
+		width: 46px;
+		height: 51px;
+		line-height: 14px;
+		padding-top: 12px;
+		text-align: center;
+		font-weight: bold;
 	}
 }
