+++ src/styles/turmoil.less
@@ -74,7 +74,7 @@
     display: inline-block;
     vertical-align: top;
     transform: scale(0.88);
-    margin: -50px 0 0px -56px;
+    margin: -60px 0 -45px -60px;
 }
 
 .events-board {
